requirements.txt
aiogram==2.3
python-decouple==3.1
mysqlclient==1.4.6
